fix: the graph ipfs storage handling (#160)
This bumps up the default config package to use the correct
The Graph IPFS API url. That url is not the same in public environments
(mumbai, ropsten, etc.).

This also fixes the CORS error that was resulting from the above, that was cause
do to incorrect handling of headers when connecting to the IPFS node of the graph.